Refactored cohort summary with single top level component and other components in child folders.
