+++ qa-portal-angular/projects/self-reflection/src/app/cohort-summary/cohort-chart/cohort-chart.component.html
