+++ qa-portal-angular/projects/self-reflection/src/app/cohort-summary/cohort-table/cohort-table.component.css
