+++ qa-portal-angular/projects/self-reflection/src/app/qa-self-reflection.module.ts
@@ -7,9 +7,9 @@ import {SummaryService} from './cohort-summary/services/summary.service';
 import {TraineeReflectionComponent} from './trainee-reflection/trainee-reflection.component';
 import {SelfReflectionHistoryComponent} from './self-reflection-history/self-reflection-history.component';
 import {CohortSummaryComponent} from './cohort-summary/cohort-summary.component';
-import {CohortTableComponent} from './cohort-table/cohort-table.component';
+import {CohortTableComponent} from './cohort-summary/cohort-table/cohort-table.component';
 import {MatTableModule, MatPaginatorModule, MatSortModule, MatRippleModule} from '@angular/material';
-import {CohortChartComponent} from './cohort-chart/cohort-chart.component';
+import {CohortChartComponent} from './cohort-summary/cohort-chart/cohort-chart.component';
 import {CohortTraineesComponent} from './cohort-trainees/cohort-trainees.component';
 import {CohortTraineesService} from './cohort-trainees/services/cohort-trainees.service';
 import {SelfReflectionFormService} from './trainee-reflection/services/self-reflection-form.service';
